do refactoring. Split note fragment on create and edit fragments with simple viewModels

+++ app/src/main/java/com/example/teachnotes/databases/Note.kt
@@ -11,7 +11,7 @@ data class Note(
 
     @PrimaryKey(autoGenerate = true)
     @ColumnInfo(name = "note_id")
-    var noteId: Int,
+    var noteId: Long,
 
     @ColumnInfo(name = "note_title")
     var noteTitle: String,
@@ -21,4 +21,5 @@ data class Note(
 
     @ColumnInfo(name = "note_is_favorite")
     var isFavorite: Boolean
+
 ) : Parcelable
